walletpage: add TransactionsTable
